+++ restapi/models/ConfirmationModel.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Table, Column, String, Integer, Boolean, ForeignKey
+from sqlalchemy import Table, Column, String, Integer, BigInteger, Boolean, ForeignKey
 from sqlalchemy.sql import expression
 from config import metadata
 
@@ -6,5 +6,5 @@ confirmation = Table('confirmation_users', metadata,
     Column('id', String(100), primary_key=True),
     Column('activated', Boolean, server_default=expression.false()),
     Column('resend_expired', Integer, nullable=True),
-    Column('user_id', Integer, ForeignKey('users.id',onupdate='cascade',ondelete='cascade'), nullable=False)
+    Column('user_id', BigInteger, ForeignKey('users.id',onupdate='cascade',ondelete='cascade'), nullable=False)
 )
